backend/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
pymongo==4.5.0
python-dotenv==1.0.0
Flask-Cors==4.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
bcrypt==4.0.1